Замена int на BigDecimal

+++ api/src/main/java/ru/gb/market/api/cart/CartDto.java
@@ -2,15 +2,16 @@ package ru.gb.market.api.cart;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
 import java.util.List;
 @Schema(description = "Модель корзины")
 public class CartDto {
     @Schema(description = "Список продуктов в корзине", required = true)
     private  List<CartItemDto> items;
     @Schema(description = "Общая стоимость продуктов в корзине", required = true)
-    private int totalPrice;
+    private BigDecimal totalPrice;
 
-    public CartDto(List<CartItemDto> items, int totalPrice) {
+    public CartDto(List<CartItemDto> items, BigDecimal totalPrice) {
         this.items = items;
         this.totalPrice = totalPrice;
     }
@@ -23,11 +24,11 @@ public class CartDto {
         this.items = items;
     }
 
-    public int getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
