+++ order-service/src/main/java/ru/gb/market/order/entities/OrderItem.java
@@ -8,6 +8,7 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -28,9 +29,9 @@ public class OrderItem {
     @Column(name = "count")
     private int count;
     @Column(name = "price_per_product")
-    private int pricePerProduct;
+    private BigDecimal pricePerProduct;
     @Column(name = "price")
-    private int price;
+    private BigDecimal price;
     @ManyToOne
     @JoinColumn(name = "order_id")
     private Order order;
